pyproject.toml
[tool.poetry]
name = "django-project"
version = "0.1.0"
description = ""
authors = ["Eason_chiang <95297596+class83108@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "4.2"
load-dotenv = "^0.1.0"
psycopg2-binary = "^2.9.9"
django-allauth = "^64.2.1"
pillow = "^10.4.0"
django-jazzmin = "^3.0.0"
django-mdeditor = "^0.1.20"
django-debug-toolbar = "4.2.0"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
